refactor(folders): shift to chakra
